Fixing tests, segregating mocks

Fixing tests and mocks
